日ノ本アップデート関連、マップデータを更新 refs #131 r70 spr_action テーブルの不具合を修正 refs #132
git-svn-id: http://svn.l1j.org/l1j/trunk@72 698199e4-35c4-443b-82c4-7f4b4ae73c13

+++ db/japanese/spr_action.sql
@@ -5,9 +5,9 @@ SET FOREIGN_KEY_CHECKS=0;
 DROP TABLE IF EXISTS `spr_action`;
 CREATE TABLE IF NOT EXISTS `spr_action` (
   `spr_id` int(10) unsigned NOT NULL,
-  `act_id` tinyint(3) unsigned NOT NULL,
-  `framecount` tinyint(3) unsigned NOT NULL DEFAULT '0',
-  `framerate` tinyint(3) unsigned NOT NULL DEFAULT '24',
+  `act_id` int(10) unsigned NOT NULL,
+  `framecount` int(10) unsigned NOT NULL DEFAULT '0',
+  `framerate` int(10) unsigned NOT NULL DEFAULT '24',
   PRIMARY KEY (`spr_id`,`act_id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
